Learning service lifetimes
