+++ application/ITimeProvider.cs
@@ -1,0 +1,6 @@
+namespace e08.application;
+
+public interface ITimeProvider
+{
+    void PrintTime();
+}
